moved all namespaces + splitted query
