+++ OdoriRails/OdoriRails/BaseClasses/Service.cs
@@ -1,15 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace OdoriRails
+namespace OdoriRails.BaseClasses
 {
     public abstract class Service
     {
-        public int Id { get; protected set; }
+        public int Id { get; private set; }
         public List<User> AssignedUsers { get; protected set; }  
-        public DateTime StartDate { get; protected set; }
-        public DateTime EndDate { get; protected set; }
-        public int TramId { get; protected set; }
+        public DateTime StartDate { get; private set; }
+        public DateTime EndDate { get; private set; }
+        public int TramId { get; private set; }
 
         protected Service(int? id, List<User> assignedUsers, DateTime startDate, DateTime? endDate, int? tramId)
         {
