+++ OdoriRails/OdoriRails/BaseClasses/Tram.cs
@@ -1,4 +1,4 @@
-﻿namespace OdoriRails
+﻿namespace OdoriRails.BaseClasses
 {
     public enum TramStatus
     {
