+++ OdoriRails/OdoriRails/DAL/MSSQLDatabaseContext.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
 using System.Linq;
+using OdoriRails.BaseClasses;
 
-
-namespace OdoriRails
+namespace OdoriRails.DAL
 {
     /// <summary>
     /// Database Adapter voor de mssql datatabase.
@@ -261,7 +261,7 @@ WHERE ([User].Username = @usrname)) AS derivedtbl_1 ON Service.ServicePk = deriv
             return repair;
         }
 
-        public List<Service> GetAllServicesWithoutUsers()
+        public List<Repair> GetAllRepairsWithoutUsers()
         {
             var repairQuery = new SqlCommand(@"SELECT Repair.*
 FROM Repair INNER JOIN
@@ -269,22 +269,21 @@ FROM Repair INNER JOIN
 FROM ServiceUser RIGHT OUTER JOIN
 Service ON ServiceUser.ServiceCk = Service.ServicePk
 WHERE (ServiceUser.UserCk IS NULL)) AS derivedtbl_1 ON Repair.ServiceFk = derivedtbl_1.ServicePk");
+            var repairData = GetData(repairQuery);
+            return GenerateListWithFunction(repairData, CreateRepair);
 
+        }
+
+        public List<Cleaning> GetAllCleansWithoutUsers()
+        {
             var cleanQuery = new SqlCommand(@"SELECT Clean.*
 FROM Clean INNER JOIN
 (SELECT Service.ServicePk
 FROM ServiceUser RIGHT OUTER JOIN
 Service ON ServiceUser.ServiceCk = Service.ServicePk
 WHERE (ServiceUser.UserCk IS NULL)) AS derivedtbl_1 ON Clean.ServiceFk = derivedtbl_1.ServicePk");
-            var repairData = GetData(repairQuery);
             var cleanData = GetData(cleanQuery);
-
-            var returnList = new List<Service>();
-
-            returnList.AddRange(GenerateListWithFunction(repairData, CreateRepair));
-            returnList.AddRange(GenerateListWithFunction(cleanData, CreateCleaning));
-
-            return returnList;
+            return GenerateListWithFunction(cleanData, CreateCleaning);
         }
